lib: nrf_modem: Increase modem lib heap size when traces are enabled
This increase in modem lib heap size is because the modem traces use
more heap size than before. If the modem lib heap fills up, socket
data will be affected and might stop. In an upcoming libmodem release
the heap needed for modem traces will use a dedicated heap, and this
change can be reversed.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>